requirements.txt
annotated-types==0.6.0
anyio==4.2.0
astroid==3.0.3
autopep8==2.0.4
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
distro==1.9.0
Flask==3.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonify==0.5
MarkupSafe==2.1.3
mccabe==0.7.0
openai==1.7.2
packaging==23.2
platformdirs==4.2.0
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pylint==3.0.3
python-decouple==3.8
requests==2.31.0
sniffio==1.3.0
tomlkit==0.12.3
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
